:art: adding fk scripts for mitchells and publications models.
