+++ lib_metadata_db/newspapers/scripts/publications_fk.py
@@ -2,30 +2,30 @@ import pandas as pd
 from joblib import Parallel, delayed
 from tqdm import tqdm
 
-from lib_metadata_db.press_directories.models.mitchells import Publication
 from lib_metadata_db.newspapers.models.mitchells_publication import MitchellsPublication
+from lib_metadata_db.newspapers.models.publications import Publication
 
 mitchells_publication_mappings = pd.read_csv(
     "lib_metadata_db/newspapers/import_files/newspapers_overview_with_links.csv", low_memory=False
 )
 
 
-def update_mitchells_publications_keys(mitchells):
-    associated_mitchells = mitchells_publication_mappings.loc[
-        mitchells_publication_mappings["link_to_mpd"]
-        == mitchells.title,
-        "S-TITLE",
+def update_publication_keys(mitchells_publication):
+    associated_mitchells_publications = mitchells_publication_mappings.loc[
+        mitchells_publication_mappings["AcquiredYears"]
+        == mitchells_publication.year,
+        "link_to_mpd",
     ]
-    MitchellsPublication.objects.filter(year__in=associated_mitchells).update(
-        place_of_publication_id=mitchells.id
+    Publication.objects.filter(publication_code__in=associated_mitchells_publications).update(
+        mitchells_publication_id=mitchells_publication.id
     )
 
 
-def mitchells_publication_foreign_key_mappings():
+def publication_foreign_key_mappings():
     Parallel(n_jobs=-1, prefer="threads", verbose=5)(
-        delayed(update_mitchells_publications_keys)(mitchells) for mitchells in tqdm(Publication.objects.all())
+        delayed(update_publication_keys)(mitchells_publication) for mitchells_publication in tqdm(MitchellsPublication.objects.all())
     )
 
 
 if __name__ == "__main__":
-    mitchells_publication_foreign_key_mappings()
+    publication_foreign_key_mappings()
